Split address model into a few sub models to support C.18(The temp address would be GeocodedAddress, but will not be categoried  as CustomerCommonAddress nor ProviderCommonAddress)
