+++ app/views/addresses/_results.html.erb
@@ -2,7 +2,7 @@
   <% for address in @addresses %>
     <tr>
       <td class="name">
-        <%=  link_to_if can?(:edit, Address), address.name || '' , edit_address_path(address) %>
+        <%=  link_to_if can?(:edit, ProviderCommonAddress), address.name || '' , edit_provider_common_address_path(address) %>
       </td>
       <td><%= address.building_name %></td>
       <td><%= address.address %></td>
@@ -12,7 +12,7 @@
       <td><%= address.in_district? ? "In District" : "Out of District" %></td>
       <td><%= address.notes %></td>
       <% if can? :edit, Address %>
-      <td><%= link_to translate_helper("edit"), edit_address_path(address), :class=>'edit edit-address' %></td>
+      <td><%= link_to translate_helper("edit"), edit_provider_common_address_path(address), :class=>'edit edit-address' %></td>
       <!-- <td><%#= link_to "Delete", address, :confirm => 'Are you sure?', :method => :delete, :class=>'delete' %></td> -->
       <% end %>
     </tr>
