+++ app/views/addresses/edit.html.erb
@@ -45,7 +45,7 @@
         <ol class="section">
           <li>
           <%= label_tag translate_helper("address_search") %>
-          <%= autocomplete_field_tag 'address_term', '', autocomplete_addresses_path(exclude:@address.try(:id)), :size => 30, :delay => 500, :id_element => "#address_id" %>
+          <%= autocomplete_field_tag 'address_term', '', autocomplete_provider_common_addresses_path(exclude:@address.try(:id)), :size => 30, :delay => 500, :id_element => "#address_id" %>
           <%= hidden_field_tag 'address_id' %>
           </li>
         </ol>
