+++ app/views/drivers/_address_panel.html.erb
@@ -2,7 +2,7 @@
 <div class="panel panel-primary">
   <div class="panel-heading"><%= translate_helper("driver_address_heading") %></div>
   <div class="panel-body form-horizontal">
-    <%= f.fields_for :driver_address, @driver.driver_address do |address_form| %>
+    <%= f.fields_for :address, @driver.address do |address_form| %>
       <%= render 'addresses/non_geocoding_form_fields', address_form: address_form, readonly: @readonly%>
     <% end %>
   </div><!-- END .panel-body -->
