+++ app/worker/address_upload_worker.rb
@@ -5,7 +5,7 @@ class AddressUploadWorker
     Rails.logger.info "AddressUploadWorker#perform, url=#{filename}"
     provider = Provider.find_by_id(provider_id)
     begin
-      Address.load_addresses(filename, provider)
+      ProviderCommonAddress.load_addresses(filename, provider)
     rescue Exception => ex
       puts ex.message
     end
