+++ spec/factories/addresses.rb
@@ -6,4 +6,12 @@ FactoryGirl.define do
     city { Faker::Address.city }
     state "OR"
   end
+
+  factory :user_address, parent: :address, class: UserAddress
+  
+  factory :driver_address, parent: :address, class: DriverAddress
+
+  factory :customer_common_address, parent: :address, class: CustomerCommonAddress
+
+  factory :provider_common_address, parent: :address, class: ProviderCommonAddress
 end
