+++ spec/factories/drivers.rb
@@ -5,7 +5,7 @@ FactoryGirl.define do
     name { Faker::Lorem.words(2).join(' ') }
     provider
     user
-    address
+    association :address, factory: :driver_address
     phone_number { Faker::PhoneNumber.phone_number }
   end
 end
