+++ spec/routing/addresses_routing_spec.rb
@@ -15,28 +15,24 @@ RSpec.describe AddressesController, type: :routing do
       expect(:get => "/addresses/1").to_not route_to("addresses#show", :id => "1")
     end
 
-    it "routes to #edit" do
-      expect(:get => "/addresses/1/edit").to route_to("addresses#edit", :id => "1")
+    it "does not route to #edit" do
+      expect(:get => "/addresses/1/edit").to_not route_to("addresses#edit", :id => "1")
     end
 
-    it "routes to #create" do
-      expect(:post => "/addresses").to route_to("addresses#create")
+    it "does not route to #create" do
+      expect(:post => "/addresses").to_not route_to("addresses#create")
     end
 
-    it "routes to #update" do
-      expect(:put => "/addresses/1").to route_to("addresses#update", :id => "1")
+    it "does not route to #update" do
+      expect(:put => "/addresses/1").to_not route_to("addresses#update", :id => "1")
     end
 
-    it "routes to #destroy" do
-      expect(:delete => "/addresses/1").to route_to("addresses#destroy", :id => "1")
+    it "does not route to #destroy" do
+      expect(:delete => "/addresses/1").to_not route_to("addresses#destroy", :id => "1")
     end
 
-    it "routes to #autocomplete" do
-      expect(:get => "/addresses/autocomplete").to route_to("addresses#autocomplete")
-    end
-
-    it "routes to #search" do
-      expect(:get => "/addresses/search").to route_to("addresses#search")
+    it "routes to #trippable_autocomplete" do
+      expect(:get => "/addresses/trippable_autocomplete").to route_to("addresses#trippable_autocomplete")
     end
 
   end
